savepoint: pattern for BaseCmdStack
