+++ .gitignore
@@ -11,6 +11,9 @@
 /db/*.sqlite3
 /db/*.sqlite3-*
 
+# Ignore postgres database.
+/db/*/postgres/downloads
+
 # Ignore mise databases
 /.mise/services/postgresql/data
