+++ lib/lar_city/cli/output_helpers.rb
@@ -39,6 +39,7 @@ module LarCity
 
         # Check if thor option exists in base context
         base.include SayHelperMethods
+        base.include FormatHelperMethods
       end
 
       module SayHelperMethods
@@ -97,6 +98,35 @@ module LarCity
 
         alias pretend? dry_run?
         alias debug? verbose?
+      end
+
+      module FormatHelperMethods
+        protected
+
+        def things(count, name: 'item')
+          name.pluralize(count)
+        end
+
+        def tally(collection, name)
+          return unless is_enumerable?(collection)
+
+          count = collection.count
+          "#{count} #{things(count, name:)}"
+        end
+
+        def range(collection)
+          return unless is_enumerable?(collection)
+          return unless collection.any?
+
+          count = collection.count
+          return '[1]' if count == 1
+
+          "[1-#{count}]"
+        end
+
+        def is_enumerable?(collection)
+          collection.class.ancestors.include?(Enumerable)
+        end
       end
     end
   end
